文章和活动界面下拉刷新和上拉加载的修改

+++ living/living/Model/Request/HomePage/LMHomelistequest.h
@@ -10,7 +10,7 @@
 
 @interface LMHomelistequest : FitBaseRequest
 
--(id)initWithPageIndex:(int)pageIndex andPageSize:(int)pageSize;
+-(id)initWithPageIndex:(NSInteger)pageIndex andPageSize:(NSInteger)pageSize;
